// src/user/role.enum.ts
export enum USER_TYPES {
  ADMIN = 'ADMIN',
  USER = 'USER',
  EDITOR = 'EDITOR',
}
export enum CrudActions {
  ReadAll = "ReadAll",
  ReadOne = "ReadOne",
  CreateOne = "CreateOne",
  UpdateOne = "UpdateOne",
  ReplaceOne = "ReplaceOne",
  DeleteOne = "DeleteOne",
}
export enum AccessController {
  USER_CONTROLLER = 'USER_CONTROLLER',
  TODO_CONTROLLER = 'TODO_CONTROLLER',
  PRODUCT_CONTROLLER = 'PRODUCT_CONTROLLER'
}